feat: split debt form and grid
